chore: Move outlier report in statistics
